requirements.txt
pandas
numpy
requests
yfinance
beautifulsoup4
lxml
pyarrow
torch
transformers
vaderSentiment
scikit-learn
colorama
feedparser
google-genai
tqdm
pyyaml
streamlit
plotly
